Removed an old unused system, more cleanup later
