+++ core/src/main/kotlin/ecs/systems/ai/InvestigateSystem.kt
@@ -1,6 +1,8 @@
 package ecs.systems.ai
 
+import ai.findPathFromTo
 import ai.pathfinding.TileGraph
+import ai.progressPath
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.systems.IteratingSystem
 import com.badlogic.gdx.ai.btree.Task
